requirements.txt
# Main
python-telegram-bot

# Service
black
isort
pylint
pytest
pytest-cov
